Split up menu sound patches
